hopefully prevent some overflows

+++ src/Simulation/Cache.h
@@ -104,9 +104,9 @@ template <MappingType mappingType> SC_MODULE(Cache) {
     } cachelineLookupTable;
 
     // ====================================== Precomputation ======================================
-    std::uint8_t addressOffsetBits{0};
-    std::uint8_t addressIndexBits{0};
-    std::uint8_t addressTagBits{0};
+    std::uint32_t addressOffsetBits{0};
+    std::uint32_t addressIndexBits{0};
+    std::uint32_t addressTagBits{0};
     std::uint32_t addressOffsetBitMask{0};
     std::uint32_t addressIndexBitMask{0};
     std::uint32_t addressTagBitMask{0};
@@ -133,6 +133,12 @@ template <MappingType mappingType> SC_MODULE(Cache) {
      * @returns An approximation of the amount of primitive gates within this caches
      */
     std::size_t calculateGateCount() const;
+
+    /**
+     * Adds internal signals to and from write buffer to the trace file
+     * @param[in] traceFile The trace file the signals shall be added to
+     */
+ //   void traceInternalSignals(sc_core::sc_trace_file* traceFile) const;
 
 #ifdef STRICT_INSTRUCTION_ORDER
     /**
@@ -230,7 +236,7 @@ template <MappingType mappingType> SC_MODULE(Cache) {
      * @param[in] numBytes  The amount of bytes we want to read
      * @returns the data segment just read in the lowest numBytes bytes of the uint32_t
      * */
-    std::uint32_t doRead(DecomposedAddress decomposedAddr, Cacheline & cacheline, std::uint8_t numBytes) noexcept;
+    std::uint32_t doRead(DecomposedAddress decomposedAddr, Cacheline & cacheline, std::uint32_t numBytes) noexcept;
     /**
      * Reads data from bus written to by RAM and copies it into the corresponding cacheline
      * */
@@ -245,7 +251,7 @@ template <MappingType mappingType> SC_MODULE(Cache) {
      * @param[in] numBytes The number of bytes of data to be written
      */
     void doWrite(Cacheline & cacheline, DecomposedAddress decomposedAddr, std::uint32_t data,
-                 std::uint8_t numBytes) noexcept;
+                 std::uint32_t numBytes) noexcept;
     /**
      * Passes the write request to the RAM through the write buffer
      * @param[in] cacheline The cacheline the data is taken from
